Change external_code generator test to use float instead of int32_t because it does not require including a header file to get the right type for int32_t and that include is very difficult to do reliably when possibly compiling with clang that isn't fully installed, etc.
Former-commit-id: 663b84cb20653543d4cfcfecf4a3cc8c9953363e

+++ test/generator/external_code_aottest.cpp
@@ -21,7 +21,7 @@ int main()
         }
     }
       
-    Buffer<int32_t> out(10, 10);
+    Buffer<float> out(10, 10);
     int ret_code = external_code(buf.raw_buffer(), out.raw_buffer());
 
     assert(ret_code == 0);
